requirements.txt
# Deep Learning
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0

# 시각화
tensorboard>=2.13.0
matplotlib>=3.7.0
seaborn>=0.12.0

# 데이터 처리
numpy>=1.24.0
pandas>=2.0.0
pillow>=9.5.0
opencv-python>=4.8.0

# 평가 지표
scikit-learn>=1.3.0

# 유틸리티
python-dotenv>=1.0.0
tqdm>=4.65.0
pyyaml>=6.0

# FastAPI (추후)
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.4.0
python-multipart>=0.0.6

# Streamlit (추후)
streamlit>=1.28.0
requests>=2.31.0
plotly>=5.17.0
